refactor: 엔티티 모델의 이름을 document로 끝나도록 이름 변경
