+++ tests/service/test_evaluation_service.py
@@ -9,8 +9,8 @@ from moview.service.evaluation_service import EvaluationService
 from moview.repository.interview_repository import InterviewRepository
 from moview.repository.question_answer.question_answer_repository import QuestionAnswerRepository
 from moview.domain.entity.interview_document import Interview
-from moview.domain.entity.question_answer.question import Question
-from moview.domain.entity.question_answer.answer import Answer
+from moview.domain.entity.question_answer.question_document import Question
+from moview.domain.entity.question_answer.answer_document import Answer
 
 
 class TestEvaluationService(asynctest.TestCase):
